% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schedule.R
\name{get_schedule}
\alias{get_schedule}
\title{Get the NFL schedule for a certain year}
\usage{
get_schedule(year = NULL, original = F)
}
\arguments{
\item{year}{Year of NFL schedule}

\item{original}{Return original data or transformed data}
}
\value{
A data.frame
}
\description{
Get the NFL schedule for a certain year
}
\examples{
get_schedule(year="2023")
}
