% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflschedule2023.R
\docType{data}
\name{nflschedule2023}
\alias{nflschedule2023}
\title{2023 NFL Schedule}
\format{
An data.frame
}
\source{
\href{https://www.4for4.com/teams/schedule/2023/grid}{4for4}
}
\usage{
data(nflschedule2023)
}
\description{
The data is the 2023 NFL schdule gleaned from
(\href{https://www.4for4.com/teams/schedule/2023/grid}{4for4})
}
\examples{
data(nflschedule2023)
}
\keyword{datasets}
