% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflschedule2023_transformed.R
\docType{data}
\name{nflschedule2023_transformed}
\alias{nflschedule2023_transformed}
\title{2023 NFL Schedule transformed}
\format{
An data.frame
}
\source{
\href{https://www.4for4.com/teams/schedule/2023/grid}{4for4}
}
\usage{
data(nflschedule2023_transformed)
}
\description{
The data is the 2023 NFL schedule gleaned from
(\href{https://www.4for4.com/teams/schedule/2023/grid}{4for4}) but it has
been transformed.
}
\examples{
data(nflschedule2023_transformed)
}
\keyword{datasets}
