% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflteams.R
\docType{data}
\name{nflteams}
\alias{nflteams}
\title{list of teams in the NFL}
\format{
An data.frame
}
\source{
\href{https://www.4for4.com/teams/schedule/2023/grid}{4for4}
}
\usage{
data(nflteams)
}
\description{
The data is the teams in the NFL
}
\examples{
data(nflteams)
}
\keyword{datasets}
