% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform a row of data from \func{get_schedule}}
\usage{
transform_data(x)
}
\arguments{
\item{x}{The data from \func{get_schedule}}
}
\value{
A data.frame.
}
\description{
Transform a row of data from \func{get_schedule}
}
